Add Bumblezone Bee Queen to Cursed Ring's blacklist (#412)
* Add Bumblezone Bee Queen to Cursed Ring's blacklist

Basically a short summary from what I said on Discord, I am the developer of Bumblezone and a large portion of the progression in my mod is locked behind trading with the Bee Queen.

However, the queen will not trade when angry and she is a Neutral Mob. This means players stuck with the Cursed Ring cannot trade with the Bee Queen and is locked out of a large portion of my mod and come to me asking why my mod is broken.

Having the Bee Queen be a default entry in the blacklist will help minimize confusion in modpacks with both of our mods on and help keep my mod playable.

I was not able to test this change in dev because Gradle complains about needing some sort of Curio dependency. But the provided ResourceLocation for the Bee Queen is correct.

* Make it hardcoded

---------

Co-authored-by: Aizistral <admin@aizistral.com>